Refactoring on room property value converters
