+++ src/test/scala/scalaomg/common/room/RoomPropertySpec.scala
@@ -4,7 +4,7 @@ import org.scalatest.BeforeAndAfter
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
 
-import scalaomg.common.room.RoomPropertyValueConversions._
+import scalaomg.common.room.RoomPropertyValue.Conversions._
 
 class RoomPropertySpec extends AnyFlatSpec
     with Matchers
