+++ src/test/scala/scalaomg/common/room/SharedRoomSpec.scala
@@ -1,6 +1,6 @@
 package scalaomg.common.room
 
-import scalaomg.common.room.RoomPropertyValueConversions._
+import scalaomg.common.room.RoomPropertyValue.Conversions._
 import org.scalatest.BeforeAndAfter
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
